+++ spec/site_prism/all_there/recursion_checker_spec.rb
@@ -4,9 +4,12 @@ describe SitePrism::AllThere::RecursionChecker do
   let(:checker) { described_class.new(double) }
 
   before do
-    allow(checker).to receive(:current_class_all_there?).and_return(true)
-    allow(checker).to receive(:section_classes_all_there?).and_return(true)
-    allow(checker).to receive(:sections_classes_all_there?).and_return(true)
+    allow(checker)
+      .to receive_messages(
+        current_class_all_there?: true,
+        section_classes_all_there?: true,
+        sections_classes_all_there?: true
+      )
   end
 
   after { SitePrism.recursion_setting = nil }
@@ -25,12 +28,6 @@ describe SitePrism::AllThere::RecursionChecker do
 
       it 'does not perform checks on descendant sections items without recursion' do
         expect(checker).not_to receive(:sections_classes_all_there?)
-
-        checker.all_there?(recursion: :none)
-      end
-
-      it 'always checks the value of SitePrism.recursion_setting' do
-        expect(SitePrism).to receive(:recursion_setting)
 
         checker.all_there?(recursion: :none)
       end
@@ -60,25 +57,49 @@ describe SitePrism::AllThere::RecursionChecker do
 
     context 'with pages that are missing some `expected_items` and also missing some descendent items' do
       before do
-        allow(checker).to receive(:current_class_all_there?).and_return(false)
-        allow(checker).to receive(:section_classes_all_there?).and_return(false)
-        allow(checker).to receive(:sections_classes_all_there?).and_return(false)
+        allow(checker)
+          .to receive_messages(
+            current_class_all_there?: false,
+            section_classes_all_there?: false,
+            sections_classes_all_there?: false
+          )
       end
 
       it 'returns `false` when checking all items without recursion' do
         expect(checker.all_there?(recursion: :none)).to be false
       end
 
+      it 'does not perform checks on descendant section items without recursion' do
+        expect(checker).not_to receive(:section_classes_all_there?)
+
+        checker.all_there?(recursion: :none)
+      end
+
+      it 'does not perform checks on descendant sections items without recursion' do
+        expect(checker).not_to receive(:sections_classes_all_there?)
+
+        checker.all_there?(recursion: :none)
+      end
+
       it 'returns `false` when checking all items with recursion' do
         expect(checker.all_there?(recursion: :one)).to be false
+      end
+
+      it 'performs checks on the page itself with recursion' do
+        expect(checker).to receive(:current_class_all_there?)
+
+        checker.all_there?(recursion: :one)
       end
     end
 
     context 'with pages that have all `expected_items` but are missing some descendent items' do
       before do
-        allow(checker).to receive(:current_class_all_there?).and_return(true)
-        allow(checker).to receive(:section_classes_all_there?).and_return(false)
-        allow(checker).to receive(:sections_classes_all_there?).and_return(false)
+        allow(checker)
+          .to receive_messages(
+            current_class_all_there?: true,
+            section_classes_all_there?: false,
+            sections_classes_all_there?: false
+          )
       end
 
       it 'returns `true` when checking all items without recursion' do
@@ -96,12 +117,12 @@ describe SitePrism::AllThere::RecursionChecker do
       checker.all_there?(recursion: :not_one)
     end
 
-    it 'will take the value of SitePrism.recursion_setting first' do
+    it 'will check the value of SitePrism.recursion_setting when no recursion parameter is passed in' do
       SitePrism.recursion_setting = :one
 
-      expect(SitePrism).not_to receive(:logger)
+      expect(SitePrism).to receive(:recursion_setting).once
 
-      checker.all_there?(recursion: :not_one)
+      checker.all_there?
     end
   end
 end