+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 ### Removed
 
 ### Added
+- Allow runtime options into `#all_there?` calls ([luke-hill] - [#55])
 
 ### Changed
+**Breaking Change**:
+- The gem now will implicitly use no recursion if no parameter is provided
+  - Providing no value will then look at the config parameter
+    ([luke-hill] - [#55])
+
+- **Required Ruby Version is now 2.7+** ([luke-hill] - [#55])
 
 ### Fixed
+- Erroneous log messages when using `sections` (plural) check in `#all_there?` ([luke-hill] - [#55])
 
 ## [2.0.2] - 2023-07-13
 ### Changed